[QUAD-271] Provide option to add as a project admin and display navigation bar based on user authorizations.
