+++ Quadriga/src/main/webapp/WEB-INF/tiles/layouts/skeleton_nomenu.jsp
@@ -58,7 +58,23 @@
         <div>
         <div class="navbar-collapse collapse navbar-right">
           <ul class="nav navbar-nav">
-            <tiles:insertAttribute name="navigation" />
+            		<sec:authorize access="hasRole('ROLE_QUADRIGA_USER_ADMIN') and !hasAnyRole('ROLE_QUADRIGA_USER_STANDARD', 'ROLE_QUADRIGA_USER_COLLABORATOR')">
+						<ul class="nav navbar-nav">
+							<tiles:insertAttribute name="navigation" />
+						</ul>
+				  	</sec:authorize>	
+				 	<sec:authorize access= "hasAnyRole('ROLE_QUADRIGA_USER_STANDARD', 'ROLE_QUADRIGA_USER_COLLABORATOR')">
+				 		<sec:authorize access="hasRole('ROLE_QUADRIGA_USER_ADMIN') ">
+				 			<ul class="nav navbar-nav" style="font-size:13px;">
+								<tiles:insertAttribute name="navigation" />
+							</ul>
+				 		</sec:authorize>
+				 		<sec:authorize access="!hasRole('ROLE_QUADRIGA_USER_ADMIN') ">
+				 			<ul class="nav navbar-nav">
+								<tiles:insertAttribute name="navigation" />
+							</ul>
+				 		</sec:authorize>
+				 	</sec:authorize>
           </ul>
         </div>
         <sec:authorize access="not isAuthenticated()">
